Moving some things to exceptions.py and utils.py
